refactored some card functionality out of the GameManager into a new CardManager class
