+++ app/src/test/java/com/uni/gruppenphaseandroid/managertests/GameManagerWormholeTest.java
@@ -14,6 +14,7 @@ import com.uni.gruppenphaseandroid.communication.Client;
 import com.uni.gruppenphaseandroid.communication.dto.Message;
 import com.uni.gruppenphaseandroid.communication.dto.MessageType;
 import com.uni.gruppenphaseandroid.communication.dto.WormholeSwitchPayload;
+import com.uni.gruppenphaseandroid.manager.CardManager;
 import com.uni.gruppenphaseandroid.manager.GameManager;
 import com.uni.gruppenphaseandroid.manager.VisualEffectsManager;
 import com.uni.gruppenphaseandroid.playingfield.FigureManager;
@@ -33,6 +34,7 @@ public class GameManagerWormholeTest {
     PlayingField playingField;
     View view;
     FigureManager figureManager;
+    CardManager cardManager;
     Card card;
     Client socketClient;
     ArrayList<Wormhole> wormholeList;
@@ -50,6 +52,7 @@ public class GameManagerWormholeTest {
         socketClient = mock(Client.class);
         GameManager.getInstance().setWebSocketClient(socketClient);
         figureManager = mock(FigureManager.class);
+        cardManager = mock(CardManager.class);
         wormholeList = new ArrayList<>();
         when(playingField.getWormholeList()).thenReturn(wormholeList);
 
@@ -64,7 +67,7 @@ public class GameManagerWormholeTest {
         message.setPayload(new Gson().toJson(payload));
 
         VisualEffectsManager visualEffectsManager = mock(VisualEffectsManager.class);
-        GameManager.getInstance().startGame(4, 2, lobbyID, playerID, figureManager, visualEffectsManager);
+        GameManager.getInstance().startGame(4, 2, lobbyID, playerID, figureManager, visualEffectsManager, cardManager);
       //  Assert.assertFalse(GameManager.getInstance().hasCheated());
     }
